requirements.txt
aiofiles==23.2.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.1.0
audioread==3.0.1
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
broken==0.1
broken-source==0.6.0
cachetools==5.4.0
cattrs==23.2.3
certifi==2025.6.15
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
depthflow==0.6.0
diskcache==5.6.3
distro==1.9.0
dotmap==1.3.30
executing==2.1.0
fastapi==0.112.2
fastjsonschema==2.20.0
ffmpy==0.4.0
filelock==3.15.4
fonttools==4.53.1
fqdn==1.5.1
fsspec==2024.6.1
glcontext==2.5.0
glfw==2.7.0
gradio==4.22.0
gradio_client==0.13.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
imageio==2.35.1
imgui==2.0.0
importlib_resources==6.4.4
iniconfig==2.1.0
intervaltree==3.1.0
ipykernel==6.29.5
ipython==8.28.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.5
lazy_loader==0.4
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
midi2audio==0.1.1
mido==1.3.2
mistune==3.0.2
moderngl==5.10.0
moderngl-window==2.4.6
mpmath==1.3.0
multipledispatch==1.0.0
narwhals==1.5.5
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.26.4
numpy-quaternion==2023.0.4
openai==1.37.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opensimplex==0.4.5.1
ordered-set==4.1.0
orjson==3.10.7
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.6.0
pretty_midi==0.2.10
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
pyfiglet==1.0.2
pyFluidSynth==1.3.3
pygame==2.6.0
pyglet==2.0.16
Pygments==2.18.0
pyparsing==3.1.2
pyrr==0.10.3
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.9
python-rtmidi==1.5.8
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.2.0
rapidfuzz==3.9.5
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-cache==1.2.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.20.0
ruff==0.6.2
safetensors==0.4.3
schedule==1.2.2
scikit-image==0.24.0
scipy==1.14.0
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
setuptools==72.1.0
shaderflow==0.6.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SoundCard==0.4.3
soupsieve==2.6
stack-data==0.6.3
starlette==0.38.2
sympy==1.13.1
terminado==0.18.1
thefuzz==0.22.1
tifffile==2024.8.10
timm==1.0.8
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.0
torch==2.2.2
torchvision==0.17.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.43.3
turbopipe==1.2.1
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
url-normalize==1.4.3
urllib3==2.2.2
uv==0.2.37
uvicorn==0.30.6
validators==0.33.0
watchdog==4.0.1
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
widgetsnbextension==4.0.13
